952_SupportAccessibilityAnalysis - added accessibility analysis tests + preparation for objective c tests
