+++ Example/Tests/NetworkTests/UploaderTests/UploaderTests.swift
@@ -790,111 +790,6 @@ class UploaderTests: NetworkBaseTest {
         XCTAssertNotNil(result?.qualityAnalysis, "quality analysis field in upload result should not be nil")
     }
     
-    // MARK: - accessibility analysis
-    func test_uploadResult_accessibiltyAnalysisParsing_ShouldParseAsExpected() {
-       
-        XCTAssertNotNil(cloudinary!.config.apiSecret, "Must set api secret for this test")
-        
-        // Given
-        let expectation = self.expectation(description: "Upload with accessibility should succeed")
-        
-        let resource: TestResourceType = .borderCollie
-        let file = resource.url
-        var result: CLDUploadResult?
-        var error: NSError?
-        
-        let params = CLDUploadRequestParams()
-        params.setAccessibilityAnalysis(true)
-        
-        // When
-        cloudinary!.createUploader().signedUpload(url: file, params: params).response({ (resultRes, errorRes) in
-            result = resultRes
-            error = errorRes
-            
-            expectation.fulfill()
-        })
-        
-        waitForExpectations(timeout: timeout, handler: nil)
-
-        // Then
-        XCTAssertNil(error, "error should be nil")
-        XCTAssertNotNil(result, "result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityScore, "accessibility analysis field in upload result should not be nil")
-        
-        // accessability analysis
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.distinctColors, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.distinctEdges, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.mostIndistinctPair
-            , "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.mostIndistinctPair?[0]
-        , "accessibility analysis field in upload result should not be nil")
-    }
-    func test_explicitResult_accessibiltyAnalysisParsing_ShouldParseAsExpected() {
-       
-        XCTAssertNotNil(cloudinary!.config.apiSecret, "Must set api secret for this test")
-        
-        let expectation = self.expectation(description: "Upload should succeed")
-        let resource: TestResourceType = .borderCollie
-        let file = resource.url
-        var result: CLDUploadResult?
-        var error: NSError?
-        
-        let params = CLDUploadRequestParams()
-        params.setAccessibilityAnalysis(true)
-        cloudinary!.createUploader().signedUpload(url: file, params: params).response({ (resultRes, errorRes) in
-            result = resultRes
-            error = errorRes
-            
-            expectation.fulfill()
-        })
-        
-        waitForExpectations(timeout: timeout, handler: nil)
-
-        callForExplicitWithAccessibility(publicId: result?.publicId)
-        
-        XCTAssertNotNil(result, "result should not be nil")
-        XCTAssertNil(error, "error should be nil")
-    }
-    func callForExplicitWithAccessibility(publicId: String?) {
-
-        guard let publicId = publicId else { return }
-        
-        // Given
-        let expectation = self.expectation(description: "Explicit call with accessibility should succeed")
-        
-        var result: CLDExplicitResult?
-        var error: NSError?
-        
-        // When
-        let params = CLDExplicitRequestParams()
-        params.setAccessibilityAnalysis(true)
-        cloudinary!.createManagementApi().explicit(publicId, type: "upload", params: params).response({ (resultRes, errorRes) in
-            result = resultRes
-            error = errorRes
-            
-            expectation.fulfill()
-        })
-        
-        waitForExpectations(timeout: timeout, handler: nil)
-        
-        // Then
-        XCTAssertNil(error, "error should be nil")
-        XCTAssertNotNil(result, "result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityScore, "accessibility analysis field in upload result should not be nil")
-        
-        // accessability analysis
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.distinctColors, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.distinctEdges, "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.mostIndistinctPair
-            , "accessibility analysis field in upload result should not be nil")
-        XCTAssertNotNil(result?.accessibilityAnalysis?.colorblindAccessibilityAnalysis?.mostIndistinctPair?[0]
-        , "accessibility analysis field in upload result should not be nil")
-    }
-    
     // MARK: - manual moderation
     func testManualModeration() {
